% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bonus_cards}
\alias{bonus_cards}
\title{Data for bonus cards}
\format{
\describe{
  \item{name}{Name. This corresponds to the equivalently named column
        in the \code{birds} data}
  \item{set}{Which game the bonus card is from. Currently one of
            'core', 'european' and 'oceania'}
  \item{automa}{Logical.  Is this card used when playing with the automa}
  \item{condition}{A short description of when this bonus applies}
  \item{explanatory_text}{A longer explanation of when this bonus applies}
  \item{victory_points}{The number of victory points awarded if the bonus
        conditions are met at the end of the game}
  \item{percent}{What percent of bird cards match the bonus condition?}
}
}
\source{
\url{https://boardgamegeek.com/filepage/193164/wingspan-bird-card-bonus-card-spreadsheet}
}
\usage{
bonus_cards
}
\description{
Data for bonus cards
}
\keyword{datasets}
